Redux typing.d.ts file refactored to /types
